Migrate and fix tests out of deprecated
