+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/rpa/RequestForJudgementNotificationServiceTest.java
@@ -1,14 +1,14 @@
-package uk.gov.hmcts.cmc.claimstore.deprecated.rpa;
+package uk.gov.hmcts.cmc.claimstore.rpa;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
-import uk.gov.hmcts.cmc.claimstore.deprecated.MockSpringTest;
+import uk.gov.hmcts.cmc.claimstore.BaseMockSpringTest;
 import uk.gov.hmcts.cmc.claimstore.events.ccj.CountyCourtJudgmentEvent;
-import uk.gov.hmcts.cmc.claimstore.rpa.RequestForJudgementNotificationService;
 import uk.gov.hmcts.cmc.claimstore.rpa.config.EmailProperties;
 import uk.gov.hmcts.cmc.domain.models.Claim;
 import uk.gov.hmcts.cmc.domain.models.CountyCourtJudgment;
@@ -17,6 +17,7 @@ import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaim;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleCountyCourtJudgment;
 import uk.gov.hmcts.cmc.email.EmailAttachment;
 import uk.gov.hmcts.cmc.email.EmailData;
+import uk.gov.hmcts.cmc.email.EmailService;
 
 import java.time.LocalDate;
 
@@ -31,13 +32,16 @@ import static uk.gov.hmcts.cmc.claimstore.rpa.ClaimIssuedNotificationService.JSO
 import static uk.gov.hmcts.cmc.claimstore.utils.DocumentNameUtils.buildJsonRequestForJudgementFileBaseName;
 import static uk.gov.hmcts.cmc.claimstore.utils.DocumentNameUtils.buildRequestForJudgementFileBaseName;
 
-public class RequestForJudgementNotificationServiceTest extends MockSpringTest {
+public class RequestForJudgementNotificationServiceTest extends BaseMockSpringTest {
     private static final byte[] PDF_CONTENT = {1, 2, 3, 4};
 
     @Autowired
     private RequestForJudgementNotificationService service;
     @Autowired
     private EmailProperties emailProperties;
+
+    @MockBean
+    protected EmailService emailService;
 
     @Captor
     private ArgumentCaptor<String> senderArgument;
