+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/services/staff/RejectSettlementAgreementStaffNotificationServiceTest.java
@@ -1,16 +1,17 @@
-package uk.gov.hmcts.cmc.claimstore.deprecated.services.staff;
+package uk.gov.hmcts.cmc.claimstore.services.staff;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import uk.gov.hmcts.cmc.claimstore.BaseMockSpringTest;
 import uk.gov.hmcts.cmc.claimstore.config.properties.emails.StaffEmailProperties;
-import uk.gov.hmcts.cmc.claimstore.deprecated.MockSpringTest;
-import uk.gov.hmcts.cmc.claimstore.services.staff.RejectSettlementAgreementStaffNotificationService;
 import uk.gov.hmcts.cmc.domain.models.Claim;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaim;
 import uk.gov.hmcts.cmc.email.EmailData;
+import uk.gov.hmcts.cmc.email.EmailService;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -18,16 +19,18 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 
-public class RejectSettlementAgreementStaffNotificationServiceTest extends MockSpringTest {
+public class RejectSettlementAgreementStaffNotificationServiceTest extends BaseMockSpringTest {
 
     @Autowired
     private RejectSettlementAgreementStaffNotificationService service;
+    @Autowired
+    private StaffEmailProperties emailProperties;
 
     @Captor
     private ArgumentCaptor<EmailData> emailDataArgument;
 
-    @Autowired
-    private StaffEmailProperties emailProperties;
+    @MockBean
+    protected EmailService emailService;
 
     private Claim claim;
