+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/services/staff/SettlementReachedStaffNotificationServiceTest.java
@@ -1,56 +1,75 @@
-package uk.gov.hmcts.cmc.claimstore.deprecated.services.staff;
+package uk.gov.hmcts.cmc.claimstore.services.staff;
 
-import org.apache.commons.io.IOUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import uk.gov.hmcts.cmc.claimstore.BaseMockSpringTest;
 import uk.gov.hmcts.cmc.claimstore.config.properties.emails.StaffEmailProperties;
-import uk.gov.hmcts.cmc.claimstore.deprecated.MockSpringTest;
-import uk.gov.hmcts.cmc.claimstore.services.staff.StatesPaidStaffNotificationService;
 import uk.gov.hmcts.cmc.domain.models.Claim;
+import uk.gov.hmcts.cmc.domain.models.offers.MadeBy;
+import uk.gov.hmcts.cmc.domain.models.offers.Settlement;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaim;
+import uk.gov.hmcts.cmc.domain.models.sampledata.SampleResponse;
+import uk.gov.hmcts.cmc.domain.models.sampledata.offers.SampleOffer;
 import uk.gov.hmcts.cmc.email.EmailAttachment;
 import uk.gov.hmcts.cmc.email.EmailData;
+import uk.gov.hmcts.cmc.email.EmailService;
 
 import java.io.IOException;
+import java.time.LocalDateTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.cmc.claimstore.documents.output.PDF.EXTENSION;
-import static uk.gov.hmcts.cmc.claimstore.utils.DocumentNameUtils.buildResponseFileBaseName;
 
-public class StatesPaidStaffNotificationServiceTest extends MockSpringTest {
+public class SettlementReachedStaffNotificationServiceTest extends BaseMockSpringTest {
+
     private static final byte[] PDF_CONTENT = {1, 2, 3, 4};
-    private Claim claimWithFullDefenceResponse;
+
+    @Autowired
+    private SettlementReachedStaffNotificationService service;
+    @Autowired
+    private StaffEmailProperties emailProperties;
 
     @Captor
     private ArgumentCaptor<String> senderArgument;
     @Captor
     private ArgumentCaptor<EmailData> emailDataArgument;
 
-    @Autowired
-    private StaffEmailProperties emailProperties;
+    @MockBean
+    protected EmailService emailService;
 
-    @Autowired
-    private StatesPaidStaffNotificationService service;
+    private Claim claim;
 
     @Before
-    public void beforeEachTest() {
-        claimWithFullDefenceResponse = SampleClaim.getClaimFullDefenceStatesPaidWithAcceptation();
+    public void setup() {
+        Settlement settlement = new Settlement();
+        settlement.makeOffer(SampleOffer.builder().build(), MadeBy.DEFENDANT, null);
+        settlement.accept(MadeBy.CLAIMANT, null);
 
+        claim = SampleClaim
+            .builder()
+            .withSettlementReachedAt(LocalDateTime.now())
+            .withResponse(SampleResponse.validDefaults())
+            .withSettlement(settlement)
+            .build();
         when(pdfServiceClient.generateFromHtml(any(byte[].class), anyMap()))
             .thenReturn(PDF_CONTENT);
     }
 
+    @Test(expected = NullPointerException.class)
+    public void shouldThrowNullPointerWhenGivenNullClaim() {
+        service.notifySettlementReached(null);
+    }
+
     @Test
     public void shouldSendEmailToExpectedRecipient() {
-
-        service.notifyStaffClaimantResponseStatesPaidSubmittedFor(claimWithFullDefenceResponse);
+        service.notifySettlementReached(claim);
 
         verify(emailService).sendEmail(senderArgument.capture(), emailDataArgument.capture());
 
@@ -58,30 +77,22 @@ public class StatesPaidStaffNotificationServiceTest extends MockSpringTest {
     }
 
     @Test
-    public void shouldSendEmailWithExpectedContentStatesPaidByFullDefence() {
-
-        service.notifyStaffClaimantResponseStatesPaidSubmittedFor(claimWithFullDefenceResponse);
+    public void shouldSendEmailWithExpectedContent() {
+        service.notifySettlementReached(claim);
 
         verify(emailService).sendEmail(senderArgument.capture(), emailDataArgument.capture());
 
-        String subject = String.format("Civil Money Claim States Paid Claimant Response: %s v %s %s",
-            claimWithFullDefenceResponse.getClaimData().getClaimant().getName(),
-            claimWithFullDefenceResponse.getClaimData().getDefendant().getName(),
-            claimWithFullDefenceResponse.getReferenceNumber()
-        );
-
         assertThat(emailDataArgument.getValue()
-            .getSubject()).isEqualTo(subject);
+            .getSubject()).startsWith("Offer accepted");
         assertThat(emailDataArgument.getValue()
             .getMessage()).startsWith(
-            "The defendant has submitted a States Paid defence."
+            "The claimant has accepted the defendant’s offer to settle their claim"
         );
     }
 
     @Test
     public void shouldSendEmailWithExpectedPDFAttachments() throws IOException {
-
-        service.notifyStaffClaimantResponseStatesPaidSubmittedFor(claimWithFullDefenceResponse);
+        service.notifySettlementReached(claim);
 
         verify(emailService).sendEmail(senderArgument.capture(), emailDataArgument.capture());
 
@@ -89,14 +100,12 @@ public class StatesPaidStaffNotificationServiceTest extends MockSpringTest {
             .getAttachments()
             .get(0);
 
-        String expectedFileName = buildResponseFileBaseName(claimWithFullDefenceResponse.getReferenceNumber())
-            + EXTENSION;
+        String expectedFileName = String.format(
+            SettlementReachedStaffNotificationService.FILE_NAME_FORMAT,
+            claim.getReferenceNumber()
+        );
 
         assertThat(emailAttachment.getContentType()).isEqualTo("application/pdf");
         assertThat(emailAttachment.getFilename()).isEqualTo(expectedFileName);
-
-        byte[] pdfContent = IOUtils.toByteArray(emailAttachment.getData()
-            .getInputStream());
-        assertThat(pdfContent).isEqualTo(PDF_CONTENT);
     }
 }
